Create SwapFirstLastWords.c
The main function is the ChangeThePlace function, and in my humble opinion it is the one that needs explanation. To begin with, initialize the array, where we will write the final string. Next, the length of the first (first_word_length) and last (last_word_length) words is determined (it is important to remember here that the length does not correspond to the indexes in the C language strings, because the numbering starts from zero). Using variables with length, we sequentially write to a new line first the last word from the original line, then the line between the necessary words and then the first word.

IMPORTANT NOTE: the program does not change the original line, it only creates a new one. I could only come up with such a solution to the problem.